feat: delivery-based throughput/ETA + progress-first batch view
- add `deliver` stage (UPLOADED→ACKED): receiver-delivery latency + the
  source for delivery rate; surfaced in the 耗时 tab percentiles
- throughput/ETA now count down to delivery (acked = done); remaining
  includes the uploaded-but-not-delivered backlog
- BatchSummary.throughput_per_min: rolling delivery rate (0 = stalled)
- batch detail → progress-first tabs (进度 default / 数据粒 / 日志 / 耗时);
  detail tabs lazy-load, removing the open-batch jank
- progress tab = per-stage WIP (PipelineHealth) + delivery throughput/ETA
- unify "完成"=acked+deleted across summary + pipeline helpers (fixes
  acked mis-bucketed as in-flight, dropped uploading)
- dashboard: system-wide delivery throughput/ETA
- SATHOP_STUCK_AGE_HOURS setting (was a hardcoded 6h constant in two places)

+++ src/sathop/orchestrator/api/admin.py
@@ -75,7 +75,7 @@ class OrchestratorInfo(BaseModel):
     max_inflight_per_worker: int
     max_retries: int
     max_pull_failures: int
-    stuck_age_hours: int
+    stuck_age_hours: float
     dev_mode: bool
     auth_open: bool
